update Kicad files to have new schematic design
This new schematic design is nested under a new folder called
alternative board electronics. It breaks apart the control circuitry and
the led attachment points to save PCB space (and this cost).

This may become the default schematics in a future revision, and the
all-in-one board will become the alternative.